+++ app/src/main/java/com/itachi1706/cheesecakeutilities/mlkit/camera/FrameMetadata.java
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package com.itachi1706.cheesecakeutilities.Modules.BarcodeTools.ui.camera;
+package com.itachi1706.cheesecakeutilities.mlkit.camera;
 
 /** Describing a frame info. */
 public class FrameMetadata {
